requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.25
certifi==2021.10.8
charset-normalizer==2.0.11
cycler==0.11.0
Django==2.2.6
django-celery==3.2.1
django-crontab==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.3
jieba==0.42.1
Jinja2==3.0.3
kiwisolver==1.3.1
kombu==3.0.37
lxml==4.9.2
MarkupSafe==2.0.1
matplotlib==3.3.4
mysql-connector-python==8.0.27
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
prettytable==2.5.0
protobuf==3.19.6
pyecharts==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
redis==2.10.6
requests==2.27.1
scipy==1.5.4
seaborn==0.11.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
tornado==6.1
typing_extensions==4.0.1
urllib3==1.26.8
vine==1.3.0
wcwidth==0.2.5
wincertstore==0.2
wordcloud==1.9.2
xlrd==1.2.0
zipp==3.6.0
